requirements.txt
# Dev environment
pip
autopep8

# App
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
WTForms==2.3.3
Werkzeug==2.0.3
python-dotenv
SQLAlchemy

